feat: consistent affinity
Divide the Ast Nodes between "nodes that produce values with
json affinity" and "nodes that produce values with SQL affinity"
inject conversion in parse time to wrap all SQL values as JSON values
run a separate pass to remove redundant conversions

in cases that may produce either type e.g. `(a -> b)` wrap them as json
like `jsonb(a -> b)`.

This means we can depend on values being JSON when we need them to be
so that later we can compile `.length` to different SQL depending on the
runtime type, using `json_type(...)`